requirements.txt
swagger-parser
pyyaml
inflect
setuptools
wheel
twine